Fix NullPointerException / serialization problems
IntegrityCompoundTestResult is a little broken wrt serialization:
It uses a file to serialize its child results instead of serializing
them into the output stream.

This approach is kinda ok when persisting the results to disk via
XStream
in jenkins, but it fails when serializing such objects from a jenkins
slave to its master over the wire.

This especially fails because the object from which to get the directory
where
the persistence file shall be created would needed to be serialized as
well. Since this class (AbstractBuild) is not serializable at al and the
field is even transient, this will never work.

As a workaround, we now explicitly pass the build's directory over the
wire. In the master/slave scenario, the results will only be available
on the
master when master and slave share the same persistence medium and
directory structure.

A better fix would obviously be to serialize child results also into the
given output stream instead of saving them to disk. But I leave this to
the maintainer, since I do not know the reasons for not doing this in
the first place.

While doing this, I also needed to
- port off of a few deprecated methods
- port to a newer jenkins LTS version
- add a dependency to the junit plugin, which holds all the test result
  classes
- add some @Override annotations

Fixes #13 and also relates to #8